flux-server/requirements.txt
fastapi
uvicorn[standard]
torch
diffusers
transformers
accelerate
safetensors
pillow
huggingface_hub
requests
sentencepiece
protobuf
peft
charset_normalizer
